The count in ChainPortion is an int now

+++ io-mokamint-node-api/src/main/java/io/mokamint/node/api/PublicNode.java
@@ -107,7 +107,7 @@ public interface PublicNode extends Node, Whisperer {
 	 * @throws InterruptedException if the current thread is interrupted while waiting for an answer to arrive
 	 * @throws ClosedNodeException if the node is closed
 	 */
-	ChainPortion getChainPortion(long start, long count) throws DatabaseException, TimeoutException, InterruptedException, ClosedNodeException;
+	ChainPortion getChainPortion(long start, int count) throws DatabaseException, TimeoutException, InterruptedException, ClosedNodeException;
 
 	/**
 	 * Yields the block with the given hash, if it has been seen by this node.
